Added video controller for upload

+++ src/utils/cloudinary.js
@@ -17,7 +17,9 @@ const uploadOnCloudinary = async (localFilePath) => {
     });
 
     console.log("file is uploaded on cloudinary", response.url);
-    fs.unlinkSync(localFilePath)
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    } 
     return response;
   } catch (error) {
     fs.unlinkSync(localFilePath); //remove locally seaved temporary file
